Rename fixtureData -> fixtures; Refactor & simplify MAP Media Player stories
