+++ src/app/containers/CpsAssetMediaPlayer/fixtures.js
